+++ libraries/exoplayer_hls/src/test/java/androidx/media3/exoplayer/hls/playlist/DefaultHlsPlaylistTrackerTest.java
@@ -22,6 +22,7 @@ import androidx.media3.datasource.DataSource;
 import androidx.media3.datasource.DefaultHttpDataSource;
 import androidx.media3.exoplayer.source.MediaSourceEventListener;
 import androidx.media3.exoplayer.upstream.DefaultLoadErrorHandlingPolicy;
+import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy;
 import androidx.media3.test.utils.TestUtil;
 import androidx.media3.test.utils.robolectric.RobolectricUtil;
 import androidx.test.core.app.ApplicationProvider;
@@ -31,6 +32,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 import okhttp3.HttpUrl;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
@@ -356,6 +358,85 @@ public class DefaultHlsPlaylistTrackerTest {
             Uri.parse(mockWebServer.url("/multivariant.m3u8").toString()),
             /* awaitedMediaPlaylistCount= */ 2);
 
+    assertRequestUrlsCalled(httpUrls);
+    assertThat(mediaPlaylists.get(0).mediaSequence).isEqualTo(10);
+    assertThat(mediaPlaylists.get(0).segments).hasSize(4);
+    assertThat(mediaPlaylists.get(0).trailingParts).hasSize(1);
+    assertThat(mediaPlaylists.get(1).mediaSequence).isEqualTo(10);
+    assertThat(mediaPlaylists.get(1).segments).hasSize(4);
+    assertThat(mediaPlaylists.get(1).trailingParts).hasSize(2);
+  }
+
+  @Test
+  public void start_lowLatencyNotScheduleReloadForNonPrimaryPlaylist() throws Exception {
+    List<HttpUrl> httpUrls =
+        enqueueWebServerResponses(
+            new String[] {
+              "/multivariant.m3u8",
+              "/media0/playlist.m3u8",
+              "/media1/playlist.m3u8",
+              "/media1/playlist.m3u8",
+              "/media1/playlist.m3u8?_HLS_msn=14&_HLS_part=0",
+            },
+            getMockResponse(SAMPLE_M3U8_LIVE_MULTIVARIANT),
+            getMockResponse(
+                SAMPLE_M3U8_LIVE_MEDIA_CAN_BLOCK_RELOAD_LOW_LATENCY_FULL_SEGMENT_PRELOAD),
+            getMockResponse(
+                SAMPLE_M3U8_LIVE_MEDIA_CAN_BLOCK_RELOAD_LOW_LATENCY_FULL_SEGMENT_PRELOAD),
+            getMockResponse(
+                SAMPLE_M3U8_LIVE_MEDIA_CAN_BLOCK_RELOAD_LOW_LATENCY_FULL_SEGMENT_PRELOAD),
+            getMockResponse(
+                SAMPLE_M3U8_LIVE_MEDIA_CAN_BLOCK_RELOAD_LOW_LATENCY_FULL_SEGMENT_PRELOAD_NEXT));
+
+    DefaultHlsPlaylistTracker defaultHlsPlaylistTracker =
+        new DefaultHlsPlaylistTracker(
+            dataType -> new DefaultHttpDataSource.Factory().createDataSource(),
+            new DefaultLoadErrorHandlingPolicy(),
+            new DefaultHlsPlaylistParserFactory());
+    List<HlsMediaPlaylist> mediaPlaylists = new ArrayList<>();
+    AtomicInteger playlistCounter = new AtomicInteger();
+    AtomicReference<TimeoutException> primaryPlaylistChangeExceptionRef = new AtomicReference<>();
+    defaultHlsPlaylistTracker.addListener(
+        new HlsPlaylistTracker.PlaylistEventListener() {
+          @Override
+          public void onPlaylistChanged() {
+            // Upon the first call of onPlaylistChanged(), we simulate the situation that the
+            // primary playlist url changes.
+            Uri url = defaultHlsPlaylistTracker.getMultivariantPlaylist().mediaPlaylistUrls.get(1);
+            if (defaultHlsPlaylistTracker.isSnapshotValid(url)) {
+              return;
+            }
+            defaultHlsPlaylistTracker.refreshPlaylist(url);
+            try {
+              // Make sure that the playlist for the new url has been refreshed and set it as the
+              // current primary playlist, before this method returns.
+              RobolectricUtil.runMainLooperUntil(
+                  () ->
+                      defaultHlsPlaylistTracker.getPlaylistSnapshot(url, /* isForPlayback= */ true)
+                          != null);
+            } catch (TimeoutException e) {
+              primaryPlaylistChangeExceptionRef.set(e);
+            }
+          }
+
+          @Override
+          public boolean onPlaylistError(
+              Uri url, LoadErrorHandlingPolicy.LoadErrorInfo loadErrorInfo, boolean forceRetry) {
+            return false;
+          }
+        });
+
+    defaultHlsPlaylistTracker.start(
+        Uri.parse(mockWebServer.url("/multivariant.m3u8").toString()),
+        new MediaSourceEventListener.EventDispatcher(),
+        mediaPlaylist -> {
+          mediaPlaylists.add(mediaPlaylist);
+          playlistCounter.addAndGet(1);
+        });
+    RobolectricUtil.runMainLooperUntil(() -> playlistCounter.get() >= 2);
+    defaultHlsPlaylistTracker.stop();
+
+    assertThat(primaryPlaylistChangeExceptionRef.get()).isNull();
     assertRequestUrlsCalled(httpUrls);
     assertThat(mediaPlaylists.get(0).mediaSequence).isEqualTo(10);
     assertThat(mediaPlaylists.get(0).segments).hasSize(4);